# docs/swagger.yaml
basePath: /
definitions:
  models.CreateOrderRequest:
    properties:
      email:
        type: string
      full_name:
        type: string
      is_paid:
        type: boolean
      payment_id:
        type: integer
      phone:
        type: string
      qr_code:
        type: string
      schedule_id:
        type: integer
      seat_ids:
        items:
          type: integer
        type: array
      total_price:
        type: integer
      user_id:
        type: integer
    type: object
...
  models.MovieAdmin:
    properties:
      background_poster:
        type: string
      created_at:
        type: string
      director_id:
        type: integer
      duration:
        description: untuk response JSON
        type: string
      id:
        type: integer
      popularity:
        type: integer
      poster:
        type: string
      release_date:
        type: string
      synopsis:
        type: string
      title:
        type: string
      updated_at:
        type: string
    type: object
...
  models.Order:
    properties:
      created_at:
        type: string
      email:
        type: string
      full_name:
        type: string
      id:
        type: integer
      is_paid:
        type: boolean
      payment_id:
        type: integer
      phone:
        type: string
      qr_code:
        type: string
      schedule_id:
        type: integer
      total_price:
        type: integer
      updated_at:
        type: string
      user_id:
        type: integer
    type: object
...
  models.Profile:
    properties:
      created_at:
        type: string
      first_name:
        type: string
      id:
        type: integer
      last_name:
        type: string
      phone:
        type: string
      profile_picture:
        type: string
      updated_at:
        type: string
      user_id:
        type: integer
    type: object
  models.ProfileResponse:
    properties:
      created_at:
        type: string
      email:
        type: string
      first_name:
        type: string
      last_name:
        type: string
      phone:
        type: string
      profile_picture:
        type: string
      role:
        type: string
      updated_at:
        type: string
      user_id:
        type: integer
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: ticketing
  title: tickitz
  version: "1.0"
paths:
  /admin/movies:
    get:
      description: Get all movies, only accessible for admin
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.MovieAdmin'
            type: array
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
      security:
      - JWTtoken: []
      summary: Get all movies (Admin)
      tags:
      - Admin
  /admin/movies/{id}:
    delete:
      description: Delete movie by ID, only accessible for admin
      parameters:
      - description: Movie ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - JWTtoken: []
      summary: Delete a movie (Admin)
      tags:
      - Admin
    patch:
      consumes:
      - multipart/form-data
      description: Edit movie data (partial update, with optional poster upload)
      parameters:
      - description: Movie ID
        in: path
        name: id
        required: true
        type: integer
      - description: Title
        in: formData
        name: title
        type: string
      - description: Director ID
        in: formData
        name: director_id
        type: integer
      - description: Poster file
        in: formData
        name: poster
        type: file
      - description: Background Poster file
        in: formData
        name: background_poster
        type: file
      - description: Release Date (YYYY-MM-DD)
        in: formData
        name: release_date
        type: string
      - description: Duration (e.g. 02:28)
        in: formData
        name: duration
        type: string
      - description: Synopsis
        in: formData
        name: synopsis
        type: string
      - description: Popularity
        in: formData
        name: popularity
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - JWTtoken: []
      summary: Patch movie (Admin)
      tags:
      - Admin
...
  /user/orders:
    post:
      consumes:
      - application/json
      description: Create an order with schedule and seats
      parameters:
      - description: Order request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.CreateOrderRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Order'
      security:
      - JWTtoken: []
      summary: Create new order
      tags:
      - orders
  /user/profile:
    get:
      description: Get profile information by user_id (join users + profile)
      parameters:
      - description: User ID
        in: query
        name: user_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ProfileResponse'
      security:
      - JWTtoken: []
      summary: Get user profile
      tags:
      - profile
    patch:
      consumes:
      - multipart/form-data
      description: Update firstname, lastname, phone, profile_picture
      parameters:
      - description: User ID
        in: query
        name: user_id
        required: true
        type: integer
      - description: First Name
        in: formData
        name: first_name
        type: string
      - description: Last Name
        in: formData
        name: last_name
        type: string
      - description: Phone
        in: formData
        name: phone
        type: string
      - description: Profile Picture
        in: formData
        name: profile_picture
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Profile'
      security:
      - JWTtoken: []
      summary: Update user profile
      tags:
      - profile
securityDefinitions:
  JWTtoken:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"